refactor: move iframe relatedt file to plugins dir
